[compile] Ajout de la hotbar >Revamp du système de poser / casser des blocks >Les blocks vont désormais se dropper >Ajout du stick >Augmentation du volume de la musique de 150% (de 0.2f a 0.5f)

+++ core/src/net/npcinteractive/TranscendanceEngine/Util/RenderUtil.java
@@ -85,7 +85,7 @@ public class RenderUtil
     @description(desc = "Draw text at specified coordinates with specified font")
     public static void DrawText(Batch batch, String text, Vector2 position, BitmapFont font)
     {
-        font.draw(batch, text, position.x, position.y);
+        font.draw(batch, text, (int)position.x, (int)position.y);
     }
 
     @description(desc = "Draw text at specified coordinates with specified font with a max width for textbox and stuff")
